# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Electronic notifications",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/api/notifications": {
            "get": {
                "description": "Возвращает все уведомления с фильтрацией по статусу и дате формирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Получить все уведомления",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус уведомления",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "начальная дата формирования",
                        "name": "formation_date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "конечная дата формирвания",
                        "name": "formation_date_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllNotificationsResponse"
                        }
                    }
                }
            }
        },
        "/api/notifications/user_confirm": {
            "put": {
                "description": "Сформировать или удалить уведомление пользователем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Сформировать уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/notifications/{notification_id}": {
            "get": {
                "description": "Возвращает подробную информацию об уведомлении и его типе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Получить одно уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.NotificationResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Позволяет изменить тип уведомления и возвращает обновлённые данные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Указать тип уведомления",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тип уведомления",
                        "name": "notification_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SwaggerUpdateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.UpdateNotificationResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет уведомление по id",
                "tags": [
                    "Уведомления"
                ],
                "summary": "Удалить уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/notifications/{notification_id}/delete_recipient/{recipient_id}": {
            "delete": {
                "description": "Удалить получателя из уведомления",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Удалить получателя из уведомления",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id получателя",
                        "name": "recipient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllRecipientsResponse"
                        }
                    }
                }
            }
        },
        "/api/notifications/{notification_id}/moderator_confirm": {
            "put": {
                "description": "Подтвердить или отменить уведомление модератором",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Подтвердить уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id уведомления",
                        "name": "notification_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.SwaggerUpdateNotificationRequest": {
            "type": "object",
            "properties": {
                "notification_type": {
                    "type": "string"
                }
            }
        },
        "ds.Recipient": {
            "type": "object",
            "required": [
                "adress",
                "age",
                "email",
                "fio"
            ],
            "properties": {
                "adress": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.AllNotificationsResponse": {
            "type": "object",
            "properties": {
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemes.NotificationOutput"
                    }
                }
            }
        },
        "schemes.AllRecipientsResponse": {
            "type": "object",
            "properties": {
                "recipients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Recipient"
                    }
                }
            }
        },
...
        "schemes.NotificationOutput": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "notification_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.NotificationResponse": {
            "type": "object",
            "properties": {
                "notification": {
                    "$ref": "#/definitions/schemes.NotificationOutput"
                },
                "recipients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Recipient"
                    }
                }
            }
        },
...
        "schemes.UpdateNotificationResponse": {
            "type": "object",
            "properties": {
                "notifications": {
                    "$ref": "#/definitions/schemes.NotificationOutput"
                }
            }
        }
    }
}